deploy7

+++ client/src/components/cart/cart.jsx
@@ -26,10 +26,10 @@ export default function Cart() {
     if (items.length > 0) {
       Swal.fire({
         title: "Esta Seguro",
-        text: "!se eliminaran todos los productos de su carrito¡",
+        text: "¡Se eliminaran todos los productos de su carrito!",
         icon: "warning",
         showCancelButton: true,
-        confirmButtonText: "Si estoy seguro",
+        confirmButtonText: "Si, estoy seguro",
         cancelButtonText: "Cancelar",
         confirmButtonColor: "#28a745",
       }).then((result) => {
@@ -41,7 +41,26 @@ export default function Cart() {
   };
 
   const removeFromCartHandler = (product) => {
-    dispatch(removeFromCart(product));
+    if (product.quantity === 1) {
+      // Mostrar el SweetAlert para confirmar la eliminación
+      Swal.fire({
+        title: "¿Estás seguro?",
+        text: "El producto se eliminará del carrito.",
+        icon: "warning",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        confirmButtonText: "Eliminar",
+        cancelButtonColor: "#08ef56",
+        confirmButtonColor: "#d33",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          dispatch(removeFromCart(product));
+        }
+      });
+    } else {
+      // Si la cantidad es mayor a 1, elimina directamente el producto
+      dispatch(removeFromCart(product));
+    }
   };
 
   const clearDetailHandler = () => {
@@ -154,7 +173,7 @@ export default function Cart() {
                   title: "¡El carrito esta vacio!",
                   text: "Por favor agregue un producto al carrito",
                   icon: "warning",
-                  confirmButtonText: "ok",
+                  confirmButtonText: "Ok",
                   confirmButtonColor: "#28a745",
                 })
               }
